write weights and joints only if there are any

+++ nel/tools/3d/gltf/libgltf/libgltf/Accessor.h
@@ -10,7 +10,7 @@ namespace gltf {
 
 struct Accessor
 {
-	uint32 bufferView;
+	size_t bufferView;
 	sint32 byteOffset;
 	ComponentType componentType;
 	size_t count;
